Fixed AoT Build
