+++ src/main.server.aot.ts
@@ -1,0 +1,4 @@
+import { startServer } from './server';
+import { ServerAppModuleNgFactory } from './modules/app/server-app.module.ngfactory';
+
+startServer(ServerAppModuleNgFactory);
